+++ src/main/java/vectorwing/farmersdelight/common/block/CookingPotBlock.java
@@ -82,7 +82,7 @@ public class CookingPotBlock extends Block implements SimpleWaterloggedBlock, En
 					if (!player.getInventory().add(servingStack)) {
 						player.drop(servingStack, false);
 					}
-					level.playSound(null, pos, ModSounds.BLOCK_TAKE_PORTION.get(), SoundSource.BLOCKS, 1.0F, 1.0F);
+					level.playSound(null, pos, ModSounds.BLOCK_FOOD_TAKE_PORTION.get(), SoundSource.BLOCKS, 1.0F, 1.0F);
 				} else {
 					player.openMenu(cookingPot, pos);
 				}